server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
mysqlclient==2.2.4
alembic==1.13.2
pydantic==2.9.2
python-dotenv==1.0.1
PyJWT==2.9.0
bcrypt==4.2.0
google-generativeai==0.7.2
httpx==0.27.2
email-validator==2.2.0
xmltodict==0.13.0
